COMP: Silence -Wsign-compare in GoogleTest EXPECT_EQ/ASSERT_EQ
GCC -Wextra reports mixed-signedness comparisons inside GoogleTest's
CmpHelperEQ template whenever EXPECT_EQ/ASSERT_EQ compares an unsigned
value against a signed integer literal. Because the comparison lives in
a template, the diagnostic is attributed to the instantiation site in
the test .cxx and is not suppressed by GoogleTest's -isystem include.

Match the literal's signedness (u suffix). In the range and aggregate
test template bodies that are instantiated for both signed and unsigned
element types, cast the literal to the operand type via
static_cast<std::decay_t<decltype(expr)>> so the fix is correct for
every instantiation instead of moving the warning to the signed one.
For the signed std::count result compared against an unsigned size,
cast the count to the unsigned type.

+++ Modules/Core/Common/test/itkOptimizerParametersGTest.cxx
@@ -32,14 +32,15 @@ TEST(OptimizerParameters, ConstructWithSpecifiedSizeAndInitialValue)
 
   for (double initialValue{ -1.0 }; initialValue <= 1.0; ++initialValue)
   {
-    EXPECT_EQ(OptimizerParametersType(0, initialValue).size(), 0);
+    EXPECT_EQ(OptimizerParametersType(0, initialValue).size(), 0u);
 
     for (size_t size{ 1 }; size <= 4; ++size)
     {
       const OptimizerParametersType optimizerParameters(size, initialValue);
 
       EXPECT_EQ(optimizerParameters.size(), size);
-      EXPECT_EQ(std::count(optimizerParameters.begin(), optimizerParameters.end(), initialValue), size);
+      EXPECT_EQ(static_cast<size_t>(std::count(optimizerParameters.begin(), optimizerParameters.end(), initialValue)),
+                size);
     }
   }
 }
